refactor: better names for remaining sync model files
